backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile ./requirements.in
#
asgiref==3.6.0
    # via
    #   django
    #   djoser
certifi==2022.12.7
    # via
    #   cloudinary
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
cloudinary==1.31.0
    # via
    #   -r ./requirements.in
    #   django-cloudinary-storage
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==39.0.0
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
dj-database-url==1.2.0
    # via django-heroku
django==3.2
    # via
    #   -r ./requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-heroku
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cloudinary-storage==0.3.0
    # via -r ./requirements.in
django-cors-headers==3.13.0
    # via -r ./requirements.in
django-heroku==0.3.1
    # via -r ./requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r ./requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r ./requirements.in
drf-writable-nested==0.7.0
    # via -r ./requirements.in
gunicorn==20.1.0
    # via -r ./requirements.in
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.2
    # via jinja2
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
pillow==9.4.0
    # via -r ./requirements.in
psycopg2==2.9.5
    # via django-heroku
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-decouple==3.7
    # via -r ./requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.7.1
    # via
    #   django
    #   djangorestframework
requests==2.28.2
    # via
    #   coreapi
    #   django-cloudinary-storage
    #   requests-oauthlib
    #   social-auth-core
    #   yarg
requests-oauthlib==1.3.1
    # via social-auth-core
shortuuid==1.0.11
    # via -r ./requirements.in
six==1.16.0
    # via
    #   cloudinary
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via
    #   -r ./requirements.in
    #   djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.14
    # via
    #   cloudinary
    #   requests
whitenoise==6.3.0
    # via
    #   -r ./requirements.in
    #   django-heroku
yarg==0.1.9
    # via -r ./requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
